End of part 4

+++ controllers/blogs.js
@@ -1,8 +1,14 @@
 const blogsRouter = require('express').Router()
 const Blog = require("../models/blog")
+const User = require("../models/user")
+const tokenExtractor = require("../utils/middleware").tokenExtractor
+const userExtractor = require("../utils/middleware").userExtractor
+
+//middleware
+//app.use(tokenExtractor)
 
 blogsRouter.get('/' , async (req, res) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user' , {blogs : 0})
     res.json(blogs)
 })
 
@@ -11,17 +17,28 @@ blogsRouter.get('/:id' , async(req , res) => {
     res.json(blog)
 })
 
-blogsRouter.post('/' , async (req, res) => {
+blogsRouter.post('/' , tokenExtractor , userExtractor , async (req, res) => {
     const body = req.body
+    const user = req.user
+    //const user = await User.findById(body.userId) //Without a token statement
+
     if(!body.likes)
         body.likes = undefined
     
     if(!body.title || !body.url)
         return res.status(400).end("You cannot omit title or url")
 
-    const blog = new Blog(body)
+    const blog = new Blog({
+        title : body.title,
+        author : user.username,
+        url : body.url,
+        likes : body.likes,
+        user : user._id
+    })
 
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
     res.status(200).json(savedBlog)
 })
 
@@ -39,10 +56,29 @@ blogsRouter.put('/:id' , async (req , res) => {
         })*/
 })
 
-blogsRouter.delete("/:id" , async (req , res) => {
+blogsRouter.delete("/:id" , tokenExtractor , userExtractor , async (req , res) => {
+    const userId = req.userId
 
+    //Validating whether its the creator of the blog or not
+    if(req.decodedToken.id !== req.userId){
+        return res.status(401).json({error : "unauthorized user"})
+    }
+
+    //Deleting the blog from the "blog list" of the creator
+    let userToBeUpdated = await User.findById(userId)
+
+    userToBeUpdated.blogs.forEach(blog => {
+        if (blog.toString() === req.params.id){
+            const res = userToBeUpdated.blogs.filter(item => item !== blog)
+            userToBeUpdated.blogs = res
+        }
+    })
+
+    await userToBeUpdated.save()
+
+    //Deleting the blog itself
     await Blog.findByIdAndRemove(req.params.id)
-    res.status(204).end()
+    res.status(204).end("Delete successful")
 
     /*Blog.findByIdAndDelete(req.params.id)
         .then(() => {
